refactor(kb-mcp,ci,scripts): rename kb-search-mcp service to kb-mcp
Drop "search" from service/artifact identifiers (dir, chart, image, package, CLI, CI wiring) while keeping Knowledge Base Search product wording.
