+++ .gitcommitizen
@@ -32,7 +32,7 @@ confluence-connector = services/confluence-connector/**
 teams-mcp = services/teams-mcp/**
 temenos-mcp = services/temenos-mcp/**
 hello-mcp = services/hello-mcp/**
-kb-search-mcp = services/kb-search-mcp/**
+kb-mcp = services/kb-mcp/**
 
 # Packages
 aes-gcm-encryption = packages/aes-gcm-encryption/**
