+++ .github/CODEOWNERS
@@ -1,5 +1,5 @@
 * @Unique-AG/Data-Flow @Unique-AG/Architects @Unique-AG/Platform
 
 /services/hello-mcp/ @Unique-AG/DS-Engels
-/services/kb-search-mcp/ @Unique-AG/DS-Engels
+/services/kb-mcp/ @Unique-AG/DS-Engels
 .release-please-manifest.json @Unique-AG/Data-Flow @Unique-AG/Architects @Unique-AG/Platform @Unique-AG/DS-Engels
