+++ .github/workflows/release-please.yaml
@@ -32,8 +32,8 @@ jobs:
       "services/outlook-semantic-mcp--version": ${{ steps.release.outputs['services/outlook-semantic-mcp--version'] }}
       "services/hello-mcp--release_created": ${{ steps.release.outputs['services/hello-mcp--release_created'] }}
       "services/hello-mcp--version": ${{ steps.release.outputs['services/hello-mcp--version'] }}
-      "services/kb-search-mcp--release_created": ${{ steps.release.outputs['services/kb-search-mcp--release_created'] }}
-      "services/kb-search-mcp--version": ${{ steps.release.outputs['services/kb-search-mcp--version'] }}
+      "services/kb-mcp--release_created": ${{ steps.release.outputs['services/kb-mcp--release_created'] }}
+      "services/kb-mcp--version": ${{ steps.release.outputs['services/kb-mcp--version'] }}
 
     steps:
       - uses: googleapis/release-please-action@45996ed1f6d02564a971a2fa1b5860e934307cf7 # v5.0.0
@@ -102,12 +102,12 @@ jobs:
       service-version: ${{ needs.release-please.outputs['services/hello-mcp--version'] }}
       dockerfile: services/hello-mcp/deploy/Dockerfile
 
-  release-kb-search-mcp:
+  release-kb-mcp:
     needs: release-please
-    if: ${{ needs.release-please.outputs['services/kb-search-mcp--release_created'] == 'true' }}
+    if: ${{ needs.release-please.outputs['services/kb-mcp--release_created'] == 'true' }}
     uses: ./.github/workflows/_template-cd.yaml
     secrets: inherit
     with:
-      service-name: kb-search-mcp
-      service-version: ${{ needs.release-please.outputs['services/kb-search-mcp--version'] }}
-      dockerfile: services/kb-search-mcp/deploy/Dockerfile
+      service-name: kb-mcp
+      service-version: ${{ needs.release-please.outputs['services/kb-mcp--version'] }}
+      dockerfile: services/kb-mcp/deploy/Dockerfile
