+++ .release-please-manifest.json
@@ -5,5 +5,5 @@
   "services/outlook-semantic-mcp": "3.2.0",
   "services/kyckr-mcp": "0.1.0",
   "services/hello-mcp": "0.1.0",
-  "services/kb-search-mcp": "0.0.1"
+  "services/kb-mcp": "0.0.1"
 }
