+++ services/kb-mcp/.env.example
@@ -1,4 +1,4 @@
-# kb-search-mcp local config.
+# kb-mcp local config.
 # cp .env.example .env  — then fill in values.
 # Replace <tenant> with your Unique tenant slug (e.g. your company env name).
 # A single .env is enough: unique-mcp / unique-toolkit / Zitadel all fall back to
@@ -12,7 +12,7 @@
 UNIQUE_MCP_LOCAL_BASE_URL=http://127.0.0.1:8003
 UNIQUE_MCP_PUBLIC_BASE_URL=http://127.0.0.1:8003
 # Deploy / tunnel example:
-# UNIQUE_MCP_PUBLIC_BASE_URL=https://kb-search-mcp.<tenant>.unique.app
+# UNIQUE_MCP_PUBLIC_BASE_URL=https://kb-mcp.<tenant>.unique.app
 
 # Optional Unique web app origin for knowledge-upload citation deep links.
 # UNIQUE_MCP_FRONTEND_BASE_URL=https://<tenant>.unique.app
