+++ services/kb-mcp/deploy/Dockerfile
@@ -5,10 +5,10 @@ FROM ghcr.io/astral-sh/uv:0.11.33-python3.12-trixie-slim@sha256:7b5a3c8f80fb9659
 ENV UV_COMPILE_BYTECODE=1 UV_LINK_MODE=copy UV_PYTHON_DOWNLOADS=0
 WORKDIR /app
 RUN --mount=type=cache,target=/root/.cache/uv \
-    --mount=type=bind,source=services/kb-search-mcp/uv.lock,target=uv.lock \
-    --mount=type=bind,source=services/kb-search-mcp/pyproject.toml,target=pyproject.toml \
+    --mount=type=bind,source=services/kb-mcp/uv.lock,target=uv.lock \
+    --mount=type=bind,source=services/kb-mcp/pyproject.toml,target=pyproject.toml \
     uv sync --frozen --no-install-project --no-dev --no-editable
-COPY services/kb-search-mcp /app
+COPY services/kb-mcp /app
 RUN --mount=type=cache,target=/root/.cache/uv \
     uv sync --frozen --no-dev --no-editable
 
@@ -20,4 +20,4 @@ COPY --from=builder --chown=app:app /app/.venv /app/.venv
 ENV PATH="/app/.venv/bin:$PATH" PYTHONUNBUFFERED=1
 USER app
 EXPOSE 8000
-CMD ["kb-search-mcp"]
+CMD ["kb-mcp"]
