+++ services/kb-mcp/deploy/helm-charts/kb-mcp/Chart.yaml
@@ -1,5 +1,5 @@
 apiVersion: v2
-name: kb-search-mcp
+name: kb-mcp
 description: Knowledge Base Search MCP — Unique KB search via FastMCP + unique-mcp.
 type: application
 version: 0.0.1
