+++ services/kb-mcp/deploy/helm-charts/kb-mcp/templates/_helpers.tpl
