+++ services/kb-mcp/deploy/helm-charts/kb-mcp/templates/all.yaml
