+++ services/kb-mcp/deploy/helm-charts/kb-mcp/values.additional.schema.json
