+++ services/kb-mcp/deploy/helm-charts/kb-mcp/values.schema.json
