+++ services/kb-mcp/deploy/helm-charts/kb-mcp/values.yaml
@@ -1,9 +1,9 @@
-# kb-search-mcp chart defaults — portable app only.
+# kb-mcp chart defaults — portable app only.
 # Unique hostname / TLS / NP-on live in the monorepo Argo overlay (Gateway API routes).
 # Third parties: helm install + expose however they want (LB, Ingress controller, Mesh…).
 
-nameOverride: kb-search-mcp
-fullnameOverride: kb-search-mcp
+nameOverride: kb-mcp
+fullnameOverride: kb-mcp
 selectorComponentLabel: server
 
 env:
@@ -20,8 +20,8 @@ env:
 #     # Public URL of THIS MCP service (Gateway/Kong host clients hit).
 #     # Used for OAuth redirect base + FastMCP Host allowlist (allowed_hosts).
 #     # Must match the routes.hostname you enable in this overlay, e.g.:
-#     #   https://kb-search-mcp.<tenant>.unique.app
-#     UNIQUE_MCP_PUBLIC_BASE_URL: https://kb-search-mcp.<tenant>.unique.app
+#     #   https://kb-mcp.<tenant>.unique.app
+#     UNIQUE_MCP_PUBLIC_BASE_URL: https://kb-mcp.<tenant>.unique.app
 #
 #     # Unique web app origin (the product UI), NOT this MCP.
 #     # Optional — only for knowledge-upload deep links in citations:
@@ -33,48 +33,48 @@ env:
 #     - name: UNIQUE_APP_KEY
 #       valueFrom:
 #         secretKeyRef:
-#           name: kb-search-mcp-unique-secret
+#           name: kb-mcp-unique-secret
 #           key: UNIQUE_APP_KEY
 #     - name: UNIQUE_APP_ID
 #       valueFrom:
 #         secretKeyRef:
-#           name: kb-search-mcp-unique-secret
+#           name: kb-mcp-unique-secret
 #           key: UNIQUE_APP_ID
 #     - name: UNIQUE_API_BASE_URL
 #       valueFrom:
 #         secretKeyRef:
-#           name: kb-search-mcp-unique-secret
+#           name: kb-mcp-unique-secret
 #           key: UNIQUE_API_BASE_URL
 #     - name: UNIQUE_APP_ENDPOINT
 #       valueFrom:
 #         secretKeyRef:
-#           name: kb-search-mcp-unique-secret
+#           name: kb-mcp-unique-secret
 #           key: UNIQUE_APP_ENDPOINT
 #     - name: UNIQUE_APP_ENDPOINT_SECRET
 #       valueFrom:
 #         secretKeyRef:
-#           name: kb-search-mcp-unique-secret
+#           name: kb-mcp-unique-secret
 #           key: UNIQUE_APP_ENDPOINT_SECRET
 #     - name: ZITADEL_BASE_URL
 #       valueFrom:
 #         secretKeyRef:
-#           name: kb-search-mcp-zitadel-secret
+#           name: kb-mcp-zitadel-secret
 #           key: ZITADEL_BASE_URL
 #     - name: ZITADEL_CLIENT_ID
 #       valueFrom:
 #         secretKeyRef:
-#           name: kb-search-mcp-zitadel-secret
+#           name: kb-mcp-zitadel-secret
 #           key: ZITADEL_CLIENT_ID
 #     - name: ZITADEL_CLIENT_SECRET
 #       valueFrom:
 #         secretKeyRef:
-#           name: kb-search-mcp-zitadel-secret
+#           name: kb-mcp-zitadel-secret
 #           key: ZITADEL_CLIENT_SECRET
 envVars: []
 
 image:
   registry: ghcr.io
-  repository: unique-ag/connectors/services/kb-search-mcp
+  repository: unique-ag/connectors/services/kb-mcp
   tag: 0.0.1 # x-release-please-version
   pullPolicy: IfNotPresent
 
@@ -144,7 +144,7 @@ internalServices:
 # Gateway API (HTTPRoute) via base chart — off until overlay sets hostname + paths.
 # Unique QA example overlay:
 #   routes:
-#     hostname: kb-search-mcp.<tenant>.unique.app
+#     hostname: kb-mcp.<tenant>.unique.app
 #     auth:
 #       jwt: false
 #     paths:
@@ -180,7 +180,7 @@ networkPolicy:
 # - UNIQUE_APP_ENDPOINT / UNIQUE_APP_ENDPOINT_SECRET
 # - ZITADEL_BASE_URL / ZITADEL_CLIENT_ID / ZITADEL_CLIENT_SECRET
 # Required via overlay env: (plain URLs, see comments above envVars):
-# - UNIQUE_MCP_PUBLIC_BASE_URL = https://kb-search-mcp.<tenant>.unique.app
+# - UNIQUE_MCP_PUBLIC_BASE_URL = https://kb-mcp.<tenant>.unique.app
 # Optional via overlay env::
 # - UNIQUE_MCP_FRONTEND_BASE_URL = https://<tenant>.unique.app (citation deep links)
 # Do NOT set UNIQUE_AUTH_USER_ID / UNIQUE_AUTH_COMPANY_ID in production —
