+++ services/kb-mcp/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "kb-search-mcp"
+name = "kb-mcp"
 version = "0.0.1"
 description = "Knowledge Base Search MCP — Unique KB search, content tree, and read_file tools"
 requires-python = ">=3.12"
@@ -14,17 +14,17 @@ dependencies = [
 ]
 
 [project.scripts]
-kb-search-mcp = "kb_search.main:main"
+kb-mcp = "kb_mcp.main:main"
 
 [build-system]
 requires = ["hatchling==1.31.0"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/kb_search"]
+packages = ["src/kb_mcp"]
 
 [tool.hatch.build.targets.wheel.sources]
-"src/kb_search" = "kb_search"
+"src/kb_mcp" = "kb_mcp"
 
 [dependency-groups]
 dev = [
