+++ services/kb-mcp/src/kb_mcp/references.py
