+++ services/kb-mcp/tests/test_search_tool.py
@@ -11,8 +11,8 @@ from unique_toolkit.experimental.components.internal_search import (
     KnowledgeBaseInternalSearchConfig,
 )
 
-from kb_search.config import SearchToolConfig
-from kb_search.references import (
+from kb_mcp.config import SearchToolConfig
+from kb_mcp.references import (
     CITATION_RESULT_INSTRUCTION,
     REFERENCE_FORMAT_INFORMATION,
     REFERENCE_META_KEY,
@@ -22,7 +22,7 @@ from kb_search.references import (
     reference_url,
     scope_id_from_folder_id_path,
 )
-from kb_search.tools.search import _TOOL_DESCRIPTION, search
+from kb_mcp.tools.search import _TOOL_DESCRIPTION, search
 
 
 def test_json_schema_has_service_config():
@@ -70,7 +70,7 @@ def _patch_post_processor(chunks: list):
     mock_pp = MagicMock()
     mock_pp.process = AsyncMock(return_value=chunks)
     return patch(
-        "kb_search.tools.search.InternalSearchPostProcessor.from_settings",
+        "kb_mcp.tools.search.InternalSearchPostProcessor.from_settings",
         return_value=mock_pp,
     )
 
@@ -78,7 +78,7 @@ def _patch_post_processor(chunks: list):
 def _patch_identity():
     """Per-request identity resolves in-body via unique_mcp; return a stub."""
     return patch(
-        "kb_search.tools.search.get_unique_settings_async",
+        "kb_mcp.tools.search.get_unique_settings_async",
         new=AsyncMock(return_value=MagicMock()),
     )
 
@@ -87,14 +87,14 @@ def _patch_frontend_settings(base_url: str | None = None):
     mock_settings = MagicMock()
     mock_settings.frontend_base_url_str.return_value = base_url
     return patch(
-        "kb_search.tools.search.KbSearchServerSettings",
+        "kb_mcp.tools.search.KbMcpServerSettings",
         return_value=mock_settings,
     )
 
 
 def _patch_resolve_scope_ids(mapping: dict[str, str] | None = None):
     return patch(
-        "kb_search.tools.search.resolve_scope_ids",
+        "kb_mcp.tools.search.resolve_scope_ids",
         new=AsyncMock(return_value=mapping or {}),
     )
 
@@ -109,7 +109,7 @@ async def test_search_calls_kb_service():
 
     with (
         patch(
-            "kb_search.tools.search.KnowledgeBaseInternalSearchService.from_config",
+            "kb_mcp.tools.search.KnowledgeBaseInternalSearchService.from_config",
             return_value=mock_service,
         ) as mock_from_config,
         _patch_post_processor(chunks),
@@ -140,7 +140,7 @@ async def test_search_uses_defaults_when_no_config_provided():
 
     with (
         patch(
-            "kb_search.tools.search.KnowledgeBaseInternalSearchService.from_config",
+            "kb_mcp.tools.search.KnowledgeBaseInternalSearchService.from_config",
             return_value=mock_service,
         ),
         _patch_post_processor(chunks),
@@ -161,7 +161,7 @@ async def test_search_uses_defaults_when_no_config_provided():
 async def test_search_returns_error_result_on_service_failure():
     with (
         patch(
-            "kb_search.tools.search.KnowledgeBaseInternalSearchService.from_config",
+            "kb_mcp.tools.search.KnowledgeBaseInternalSearchService.from_config",
             side_effect=RuntimeError("KB unavailable"),
         ),
         _patch_identity(),
@@ -187,11 +187,11 @@ async def test_search_returns_error_result_on_post_processor_failure():
 
     with (
         patch(
-            "kb_search.tools.search.KnowledgeBaseInternalSearchService.from_config",
+            "kb_mcp.tools.search.KnowledgeBaseInternalSearchService.from_config",
             return_value=mock_service,
         ),
         patch(
-            "kb_search.tools.search.InternalSearchPostProcessor.from_settings",
+            "kb_mcp.tools.search.InternalSearchPostProcessor.from_settings",
             return_value=mock_pp,
         ),
         _patch_identity(),
@@ -208,7 +208,7 @@ async def test_search_returns_error_result_on_post_processor_failure():
 @pytest.mark.asyncio
 async def test_search_returns_error_when_identity_unresolvable():
     with patch(
-        "kb_search.tools.search.get_unique_settings_async",
+        "kb_mcp.tools.search.get_unique_settings_async",
         new=AsyncMock(side_effect=ValueError("Refusing to fall back to UNIQUE_AUTH_")),
     ):
         result = await search(
@@ -359,7 +359,7 @@ async def test_search_results_are_numbered_sequentially_and_include_citation_blo
 
     with (
         patch(
-            "kb_search.tools.search.KnowledgeBaseInternalSearchService.from_config",
+            "kb_mcp.tools.search.KnowledgeBaseInternalSearchService.from_config",
             return_value=mock_service,
         ),
         _patch_post_processor(chunks),
@@ -392,7 +392,7 @@ async def test_search_uses_frontend_deep_links_when_scopes_resolved():
 
     with (
         patch(
-            "kb_search.tools.search.KnowledgeBaseInternalSearchService.from_config",
+            "kb_mcp.tools.search.KnowledgeBaseInternalSearchService.from_config",
             return_value=mock_service,
         ),
         _patch_post_processor(chunks),
